refactor: 관리 관련 프래그먼트 dir 수정
