+++ feature/manage/src/main/java/com/lst_1995/manage/password/ManagePasswordFragment.kt
@@ -1,9 +1,10 @@
-package com.lst_1995.manage
+package com.lst_1995.manage.password
 
 import android.os.Bundle
 import android.view.View
 import androidx.navigation.fragment.findNavController
 import com.lst_1995.core.ui.BaseFragment
+import com.lst_1995.manage.R
 import com.lst_1995.manage.databinding.FragmentManagePasswordBinding
 import dagger.hilt.android.AndroidEntryPoint
