+++ feature/manage/src/main/java/com/lst_1995/manage/table/ManageTableFragment.kt
@@ -1,0 +1,19 @@
+package com.lst_1995.manage.table
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.lst_1995.manage.R
+
+class ManageTableFragment : Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?,
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_manage_table, container, false)
+    }
+}
