+++ feature/manage/src/main/java/com/lst_1995/manage/theme/ManageThemeFragment.kt
@@ -1,18 +1,19 @@
-package com.lst_1995.manage
+package com.lst_1995.manage.theme
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.lst_1995.manage.R
 
-class ManageTableFragment : Fragment() {
+class ManageThemeFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?,
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_manage_table, container, false)
+        return inflater.inflate(R.layout.fragment_manage_theme, container, false)
     }
 }
