+++ feature/manage/src/main/java/com/lst_1995/manage/waitscreen/ManageWaitScreenFragment.kt
@@ -1,0 +1,19 @@
+package com.lst_1995.manage.waitscreen
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.lst_1995.manage.R
+
+class ManageWaitScreenFragment : Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?,
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_manage_wait_screen, container, false)
+    }
+}
