+++ feature/manage/src/main/res/layout/fragment_manage_home.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="viewModel"
-            type="com.lst_1995.manage.ManageHomeViewModel" />
+            type="com.lst_1995.manage.home.ManageHomeViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
