+++ feature/manage/src/main/res/layout/fragment_manage_password.xml
@@ -6,7 +6,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ManagePasswordFragment" >
+        tools:context=".password.ManagePasswordFragment" >
 
         <Button
             style="@style/outlinedButton"
