+++ feature/manage/src/main/res/layout/fragment_manage_table.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ManageTableFragment">
+    tools:context=".table.ManageTableFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
