+++ feature/manage/src/main/res/layout/fragment_manage_theme.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ManageThemeFragment">
+    tools:context=".theme.ManageThemeFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
