+++ feature/manage/src/main/res/layout/fragment_manage_wait_screen.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ManageWaitScreenFragment">
+    tools:context=".waitscreen.ManageWaitScreenFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
