+++ feature/manage/src/main/res/layout/fragment_sales_history.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SalesHistoryFragment">
+    tools:context=".sales_history.SalesHistoryFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
